requirements.txt
pandas>=2.0.0
numpy==1.26.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
seaborn>=0.12.0
matplotlib>=3.7.0
xgboost>=2.0.0
streamlit>=1.28.0
pyyaml>=6.0
joblib>=1.3.0
shap>=0.42.1
lime==0.2.0.1
numba>=0.57.0,<0.60
